requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
tensorflow==2.14.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
dash==2.14.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0
python-jose==3.3.0
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0
